Preliminary bugfix: type oid
With the new pg-9.6, the oid of geopgraphy and geometry changed, causing
'register_postgis2shapely' to become ineffective. I adjusted the oids
manually for now, so that this build should work with the LAV server.
Outside of that, there may be problems --> have the module detect the
correct OIDs dynamically using a connection object.