PostGIS extension disabled if shapely not installed.
The PostGIS extension auto-enables if PostGIS is installed on the remote
connection. If shapely is not installed, this will fail. But as it turns
out, many users of our database do not need the geo-spatial data. As it
would be cumbersome to teach them how to install shapely, this edit just
makes the activation process fail gracefully if shapely is not found (it
outputs a helpful message, and just keeps loading the rest of the
extension).